+++ src/Renderer/SpriteAnimator.h
@@ -10,14 +10,14 @@ namespace RenderEngine
 	{
 	public:
 		SpriteAnimator( std::shared_ptr<Sprite> pSprite );
-		void Update( uint64_t delta );
+		void Update( double delta );
 
 		size_t GetCurrentFrame() const { return currentFrame; }
 
 	private:
 		std::shared_ptr<Sprite> pSprite;
 		size_t currentFrame = 0;
-		uint64_t currentFrameDuration;
-		uint64_t currentAnimationTime = 0;
+		double currentFrameDuration;
+		double currentAnimationTime = 0.0;
 	};
 }
